+++ handlers/stripe-checkout-spike/src/services/sqs.ts
@@ -19,39 +19,8 @@ export const sendMessageToSqsQueue = async ({
 		const command = new SendMessageCommand(input);
 		const response = await client.send(command);
 		console.log(response);
-
-		// { // SendMessageResult
-		//   MD5OfMessageBody: "STRING_VALUE",
-		//   MD5OfMessageAttributes: "STRING_VALUE",
-		//   MD5OfMessageSystemAttributes: "STRING_VALUE",
-		//   MessageId: "STRING_VALUE",
-		//   SequenceNumber: "STRING_VALUE",
-		// };
 	} catch (error) {
 		console.error(error);
 		throw error;
 	}
 };
-
-// def toJsonContributorRowSqsMessage: String = {
-//     ContributorRowSqsMessage(
-//       To = ToSqsMessage(
-//         Address = email,
-//         SubscriberKey = email,
-//         ContactAttributes = ContactAttributesSqsMessage(
-//           SubscriberAttributes = SubscriberAttributesSqsMessage(
-//             EmailAddress = email,
-//             edition = edition,
-//             `payment method` = renderPaymentMethod,
-//             currency,
-//             amount = amount.setScale(2).toString,
-//             first_name = firstName,
-//             date_of_payment = formattedDate,
-//           ),
-//         ),
-//       ),
-//       DataExtensionName = "contribution-thank-you",
-//       IdentityUserId = identityId.toString,
-//     ).asJson.toString()
-//   }
-// }
